requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
python-dotenv==1.1.1
email-validator==2.1.1
httpx==0.27.0
